# tpcc: index hints and direction improvements