--- a/tpcc/order_status.go
+++ b/tpcc/order_status.go
@@ -96,7 +96,7 @@ func (o orderStatus) run(db *sql.DB, wID int) (interface{}, error) {
 				// Case 2: Pick the middle row, rounded up, from the selection by last name.
 				rows, err := tx.Query(`
 					SELECT c_id, c_balance, c_first, c_middle
-					FROM customer
+					FROM customer@customer_idx
 					WHERE c_w_id = $1 AND c_d_id = $2 AND c_last = $3
 					ORDER BY c_first ASC`,
 					wID, d.dID, d.cLast)