@jordanlewis 2017-09-21
These optimizations don't significantly change the final tpmC number in either the interleaved nor the non-interleaved case. However, they improve the throughput of the orderStatus transaction by almost an order of magnitude in both cases:

```
# Before
[11:21:1]% ./tpcc -interleave -mix=orderStatus=1 -ops-stats                                                                                                                                    jordan@Termato:~/repo/src/github.com/cockroachdb/loadgen/tpcc dtch master
_time______opName__ops/s(inst)__ops/s(cum)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
   1s         all          1.0         1.0    536.9    536.9    536.9    536.9
   1s        tpmC          0.0         0.0      0.0      0.0      0.0      0.0
   1s     payment          0.0         0.0      0.0      0.0      0.0      0.0
   1s orderStatus          1.0         1.0    536.9    536.9    536.9    536.9
   1s    delivery          0.0         0.0      0.0      0.0      0.0      0.0
   1s  stockLevel          0.0         0.0      0.0      0.0      0.0      0.0
   2s         all          9.0         5.0   1946.2   1946.2   1946.2   1946.2
   2s        tpmC          0.0         0.0      0.0      0.0      0.0      0.0
   2s     payment          0.0         0.0      0.0      0.0      0.0      0.0
   2s orderStatus          9.0         5.0   1946.2   1946.2   1946.2   1946.2
   2s    delivery          0.0         0.0      0.0      0.0      0.0      0.0
   2s  stockLevel          0.0         0.0      0.0      0.0      0.0      0.0

# After

[11:27:1]% ./tpcc -interleave -mix=orderStatus=1 -ops-stats                                                                                                                           jordan@Termato:~/repo/src/github.com/cockroachdb/loadgen/tpcc on opts -> mine/opts
_time______opName__ops/s(inst)__ops/s(cum)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
   1s         all         57.9        57.9    234.9    369.1    402.7    419.4
   1s        tpmC          0.0         0.0      0.0      0.0      0.0      0.0
   1s     payment          0.0         0.0      0.0      0.0      0.0      0.0
   1s orderStatus         57.9        57.9    234.9    369.1    402.7    419.4
   1s    delivery          0.0         0.0      0.0      0.0      0.0      0.0
   1s  stockLevel          0.0         0.0      0.0      0.0      0.0      0.0
   2s         all         74.8        66.3    243.3    385.9    436.2    469.8
   2s        tpmC          0.0         0.0      0.0      0.0      0.0      0.0
   2s     payment          0.0         0.0      0.0      0.0      0.0      0.0
   2s orderStatus         74.8        66.3    243.3    385.9    436.2    469.8
   2s    delivery          0.0         0.0      0.0      0.0      0.0      0.0
   2s  stockLevel          0.0         0.0      0.0      0.0      0.0      0.0
```


```
# Before:

[11:17]% ./tpcc -interleave -mix=orderStatus=1 -ops-stats                                                                                                                                                  jordan@Termato:~/repo/src/github.com/cockroachdb/loadgen/tpcc dtch master
_time______opName__ops/s(inst)__ops/s(cum)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
   1s         all         28.0        28.0    419.4    704.6    771.8    771.8
   1s        tpmC          0.0         0.0      0.0      0.0      0.0      0.0
   1s     payment          0.0         0.0      0.0      0.0      0.0      0.0
   1s orderStatus         28.0        28.0    419.4    704.6    771.8    771.8
   1s    delivery          0.0         0.0      0.0      0.0      0.0      0.0
   1s  stockLevel          0.0         0.0      0.0      0.0      0.0      0.0
   2s         all         36.9        32.4    436.2    671.1    671.1    671.1
   2s        tpmC          0.0         0.0      0.0      0.0      0.0      0.0
   2s     payment          0.0         0.0      0.0      0.0      0.0      0.0
   2s orderStatus         36.9        32.4    436.2    671.1    671.1    671.1
   2s    delivery          0.0         0.0      0.0      0.0      0.0      0.0
   2s  stockLevel          0.0         0.0      0.0      0.0      0.0      0.0

# After:

[11:33]% ./tpcc -mix=orderStatus=1 -ops-stats                                                                                                                                         jordan@Termato:~/repo/src/github.com/cockroachdb/loadgen/tpcc on opts -> mine/opts
_time______opName__ops/s(inst)__ops/s(cum)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
   1s         all        463.2       462.9     32.5     56.6     71.3    113.2
   1s        tpmC          0.0         0.0      0.0      0.0      0.0      0.0
   1s     payment          0.0         0.0      0.0      0.0      0.0      0.0
   1s orderStatus        463.2       462.8     32.5     56.6     71.3    113.2
   1s    delivery          0.0         0.0      0.0      0.0      0.0      0.0
   1s  stockLevel          0.0         0.0      0.0      0.0      0.0      0.0
   2s         all        459.5       461.2     33.6     65.0     79.7    121.6
   2s        tpmC          0.0         0.0      0.0      0.0      0.0      0.0
   2s     payment          0.0         0.0      0.0      0.0      0.0      0.0
   2s orderStatus        459.5       461.2     33.6     65.0     79.7    121.6
   2s    delivery          0.0         0.0      0.0      0.0      0.0      0.0
   2s  stockLevel          0.0         0.0      0.0      0.0      0.0      0.0
```