@jordanlewis 2017-09-21
Err, I accidentally added an `-interleave` to the 3rd result listing there in the text but not in the experiment. Look at the results in the web view for the correct listing.