@petermattis 2017-09-21
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Review status: 0 of 3 files reviewed at latest revision, all discussions resolved.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/loadgen/80#-:-Ku_EjPp-FSLofO6oA8S:bnfp4nl)*
<!-- Sent from Reviewable.io -->